+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Allow user to base encumbrance on "Initial encumbrance" during rollover. Refs UIF-324.
 * Edit restrictExpenditures, restrictEncumbrance ledger toggles in create/edit form. Refs UIF-349.
 * Add a return to Finance default search to app context menu dropdown. Refs UIF-350.
+* Display transactions as Voided for canceled invoices. Refs UIF-316.
 
 ## [3.0.1](https://github.com/folio-org/ui-finance/tree/v3.0.1) (2021-11-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.0.0...v3.0.1)